+++ QtViewSystem/OpenGLLayouts/opengl_linearlayout.h
@@ -21,7 +21,7 @@ public:
 
     // OpenGL_View interface
 public:
-    virtual void onMeasure(int width, int height) override;
+    virtual void onMeasure(qreal width, qreal height) override;
 };
 
 #endif // OPENGL_LINEARLAYOUT_H
